+++ CHANGELOG.md
@@ -7,6 +7,12 @@
   - Added `ValidateEntityResolutionRulesError` and `MergeEntityResolutionRulesError`.
 - Added `EntityRulesResolver`.
   - `resolveEntityResolutionRules`: returns a `EntityResolutionRulesResolved`.
+  - Added `registerContextProvider(EntityRulesContextProvider)`.
+- `APIRoot`:
+  - Initialization register: `EntityRulesResolver.registerContextProvider`.
+- Added `APIEntityResolutionRules`.
+- `APIRouteHandler`: added `entityResolutionRules`.
+- `APIRequest`: added `routeHandler`.
 - Moved entity rules classes to `bones_api_entity_rules.dart`.
 
 ## 1.3.39